Allow contextual actions on collections on a state
Sometimes a microstate contains an enumeration of related sub-objects
such as with the multiple choice and single choice states, and you want
the actual actions to be located *on* those substates so that the
context of the microstate is implicit.

So, for example, without contextual actions, in order to toggle an
option, you have to generate an action that references both the
choice *and* the option.

```
{{#with (choice) as |choice|}}
  {{#each choice.options as |option|}}
    <button onclick={{action choice.toggle option}}>Toggle</button>
  {{/each}}
{{/with}}
```

But what if instead, we just had to reference only the option and the
choice was implicit.

```
{{#with (choice) as |choice|}}
  {{#each choice.options as |option|}}
    <button onclick={{action option.toggle}}>Toggle</button>
  {{/each}}
{{/with}}
```

In that case, the `toggle` action lives right on the option, and the
`choice` is implicit since every option must by necessity be associated
with a choice. Basically this means there's less things to screw up.