state initialization is most flexible of all time
With the old system of initialization there were two ways to set the
initial state, with the `default` property or with the `initialize`
function. The former allowed you to provide a static value if there was
no value provided, the second allowed you to transform the values passed
into the helper using a function.

This replaces both of these methods with a single, more powerful
initialize() function that takes the *prior* state as its first
argument, the `params` and `options` as passed to the compute function
as the second and third arguments. This allows microstates, should they
choose, to take into account the _prior_ state in their
computation. Most microstates will not need this information, but it
would be necessary for something that wanted to have undo functionality
for example.