normalize property descriptors
The property descriptors were adding a lot of noise to the collecting of
actions for the state objects.

Rather than having to nest each action function in an object, this
introduces the `descriptor` function which creates a read-only property
descriptor for a static value, so we can specify the actions directly.

There is still more extraction to do, but this is a first start.