initialization is just another transition action
Resolves #6

This adds the concept of a `recompute` action to all microstates. When
a helper is computed because of its inputs, the recompute action is
invoked just like any other action.

Note: an interesting side effect of this change is that templates can
now bind the `recompute` action, which will have the effect of
"resetting" the state to its initial value based on the inputs.

Along with this are several enhancements:

* The `setState` method has been renamed to `transition`
* All action hashes in the entire prototype chain are merged together so
  that subclasses can inherit actions from their superclass.