allow proper coercion for number values.
Because actions can only be added to Objects, we need to use `Number`
objects and `String` objects inside the templates. However, when they
are coerced into primitives for operations like addittion and
subtraction, apparently the `Number.prototype.valueOf()` isn't good
enough (I'm not sure why).

This adds a `valueOf()` method to the Number object that wraps the
actual number which just coerces it to a primitive number as you would
expect.

Fixes #9