extract microstate functionality into pure JS
The current microstate implementation is bound to Ember Helpers, but the
fundamental flow initializing with a state, invoking an action to cause
a state transition and observing the result seems universal.

This extracts that state transition cycle into a simple object. To
accomplish this, it uses the ES Observable polyfill. To create a
Microstate with the initial state and transition properties:

  let microstate = new MicroState({value: 1}, {
    actions: {
      increment(current) {
        return {value: current.value + 1}
      }
    }
  });

In order to observe changes on this microstate, we call the subscribe
method and pass a callback which will be invoked whenever a new state
change is aviable:

  let subscription = microstate.subscribe(transition => handle(t));

The `transition` object contains four properties:

  {
   actionName,
   previous,
   next,
   state
  }

* `actionName`: the name of the action that was invoked to cause this
  state transition. E.g. `increment`
* `previous`: the previous value of the microstate
* `next`: the new value of the microstate
* `state` the object that contains the new version with the actions
  attached to it.

The inital action interface will be on the subscription, so for example

  subscription.initial.increment();

will invoke the call back with the following transition object:

  {
    actionName: 'increment',
    previous: {value: 1},
    next: {value: 2}
    state: {
      increment() { /* will invoke emit new state*/}
    }
  }

In order to unsubscribe from the observable, just invoke the unsubscribe
method on the subscription.

  subscription.unsubscribe();