Adding automated NPM publishing (#205)
## Motivation

We want to eliminate as much as possible any friction in consuming packages published from our repositories. This is important because when you find a bug or make an improvement, nothing feels worse than waiting for the package to be release. It also creates unnecessary dependency on the maintainers to manually publish.

## Approach

I'm adding GitHub Actions that Frontside developing that we can call Transparent NPM Publishing workflow. We'll eventually write a long blog post explaining how it works, but in the mean time we can enjoy the benefits of publishing without running `npm publish`.

There are 3 actions in play here,

1. [publish-pr-preview](https://github.com/thefrontside/actions/tree/master/publish-pr-preview) - automatically publish packages from pull request
2. [syncronize-with-npm](https://github.com/thefrontside/actions/tree/master/synchronize-with-npm) which runs on master and branches starting with `release-` to publish to npm whenever version in master changed
3. [synchronize-npm-tags](https://github.com/thefrontside/actions/tree/master/synchronize-npm-tags) - *I need to figure out what this does*

The workflows are as follows,

### When you want to consume changes from a PR

1. Create a pull request
2. You'll get a comment with information about package that was published

### When you want to release

1. Create a pull request with change to version in package.json
2. Once you merge, it'll automatically publish